requirements.txt
asgiref
crispy-bootstrap5
Django
django-crispy-forms
django-filter
djangorestframework
et_xmlfile
numpy
openpyxl
pandas
python-dateutil
pytz
six
sqlparse
tzdata
reportlab
arabic-reshaper
python-bidi
